+++ src/extension.ts
@@ -17,7 +17,7 @@ export async function activate(context: vscode.ExtensionContext) {
 
     if (vscode.workspace.workspaceFolders !== undefined) {
         let myStatusBarItem: vscode.StatusBarItem = vscode.window.createStatusBarItem(vscode.StatusBarAlignment.Left, 100);
-        const version = vscode.extensions.getExtension("NativeBit.i18n-l10n-editor");
+        const version = vscode.extensions.getExtension("NativeBit.i18n-l10n-editor").packageJSON.version;
         myStatusBarItem.command = 'i18n-l10n-editor';
         myStatusBarItem.text = `$(globe) i18n/l10n Editor ${version}`;
         myStatusBarItem.show();