+++ src/i18n-l10n-editor/services/translations/ije-microsoft-translator.ts
@@ -73,7 +73,6 @@ export class IJEMicrosoftTranslator implements IJETranslation {
             try {
                 var response = await axios({
                     baseURL: endpoint + `/translate`,
-                    //url: '/translate?api-version=3.0&from=en&to=fr',
                     method: 'post',
                     headers: {
                         'Ocp-Apim-Subscription-Key': apiKey,
@@ -109,9 +108,7 @@ export class IJEMicrosoftTranslator implements IJETranslation {
                                 if (to.indexOf('-') !== -1) {
                                     to = t.to.substring(0, t.to.indexOf('-'));
                                 }
-                                // if (to.indexOf('-') !== -1) {
-                                //     to = t.to.substring(0, t.to.indexOf('-'));
-                                // }
+                                
                                 if (l.indexOf(to) !== -1 || l.indexOf(t.to) !== -1) {
                                     return t.text as string;
                                 }
@@ -126,7 +123,6 @@ export class IJEMicrosoftTranslator implements IJETranslation {
                             let r = results[l][0];
                             if (r) {
                                 let _text = r['text'];
-                                //const _l = r['to'];
                                 place = 0;
                                 _substitutes.forEach(s => {
                                     if (_text.indexOf('}', place) > _text.indexOf('{', place)) {
@@ -135,10 +131,7 @@ export class IJEMicrosoftTranslator implements IJETranslation {
                                     }
                                 });
                                 translation.languages[l] = _text;
-                                //translation.languages[_l] = _text;
                                 this._manager.refreshDataTable();
-                                //const sleep = (ms: number) => new Promise(resolve => setTimeout(resolve, ms));
-                                //sleep(250);
                             }
                         }
                     });